+++ packages/app-extension/src/components/Unlocked/Settings/AddConnectWallet/ImportMnemonic.tsx
@@ -1,5 +1,4 @@
 import { useEffect, useState } from "react";
-import { mnemonicPathToPrivateKey } from "@coral-xyz/blockchain-common";
 import type {
   Blockchain,
   SignedWalletDescriptor,
@@ -19,6 +18,7 @@ import {
   useDehydratedWallets,
   useRpcRequests,
 } from "@coral-xyz/recoil";
+import { mnemonicPathToPrivateKey } from "@coral-xyz/secure-background";
 import { useCustomTheme } from "@coral-xyz/themes";
 import { Box, Typography } from "@mui/material";
