+++ packages/background/package.json
@@ -14,11 +14,9 @@
     "start": "parcel src/service-worker-loader.html -p 9333"
   },
   "dependencies": {
-    "@coral-xyz/blockchain-common": "*",
-    "@coral-xyz/blockchain-keyring": "*",
-    "@coral-xyz/secure-background": "*",
     "@coral-xyz/common": "*",
     "@coral-xyz/recoil": "*",
+    "@coral-xyz/secure-background": "*",
     "@solana/web3.js": "^1.63.1",
     "bip32": "^2.0.6",
     "bip39": "^3.0.4",
