--- packages/blockchains/common/.eslintrc.js
@@ -1,8 +1,0 @@
-module.exports = {
-  root: true,
-  extends: ["custom"],
-  parserOptions: {
-    tsconfigRootDir: __dirname,
-    project: ["./tsconfig.json"],
-  },
-};
