--- packages/blockchains/common/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../../../tsconfig.base.json",
-  "compilerOptions": {
-    "outDir": "dist/esm/",
-    "rootDir": "./src"
-  }
-}
