--- packages/blockchains/evm/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "@coral-xyz/blockchain-evm",
-  "version": "0.1.0",
-  "description": "",
-  "license": "ISC",
-  "module": "dist/esm/index.js",
-  "source": "src/index.ts",
-  "react-native": "dist/esm/index.js",
-  "types": "dist/esm/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "tsc",
-    "dev": "tsc --watch",
-    "lint": "eslint ./src --ext .js,.jsx,.ts,.tsx --cache",
-    "lint:fix": "yarn run lint --fix"
-  },
-  "dependencies": {
-    "@coral-xyz/blockchain-keyring": "*",
-    "@coral-xyz/common": "*",
-    "ethers": "^5.7.2"
-  },
-  "devDependencies": {
-    "eslint-config-custom": "*",
-    "typescript": "~4.9.3"
-  }
-}
