--- packages/blockchains/evm/src/index.ts
@@ -1,2 +1,0 @@
-export * from "./keyring";
-export * from "./util";
