--- packages/blockchains/keyring/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "@coral-xyz/blockchain-keyring",
-  "version": "0.1.0",
-  "description": "",
-  "license": "ISC",
-  "module": "dist/esm/index.js",
-  "source": "src/index.ts",
-  "react-native": "dist/esm/index.js",
-  "types": "dist/esm/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "tsc",
-    "dev": "tsc --watch",
-    "lint": "eslint ./src --ext .js,.jsx,.ts,.tsx --cache",
-    "lint:fix": "yarn run lint --fix"
-  },
-  "dependencies": {
-    "@coral-xyz/common": "*",
-    "bs58": "^5.0.0"
-  },
-  "devDependencies": {
-    "@types/node": "^18.14.6",
-    "eslint-config-custom": "*",
-    "typescript": "~4.9.3"
-  }
-}
