+++ packages/secure-background/package.json
@@ -17,11 +17,16 @@
     "lint:fix": "yarn run lint --fix"
   },
   "dependencies": {
-    "@coral-xyz/blockchain-evm": "*",
-    "@coral-xyz/blockchain-keyring": "*",
-    "@coral-xyz/blockchain-solana": "*",
     "@coral-xyz/common": "*",
-    "eventemitter3": "^5.0.1"
+    "@coral-xyz/recoil": "*",
+    "@solana/web3.js": "^1.63.1",
+    "bip32": "^2.0.6",
+    "bip39": "^3.0.4",
+    "bs58": "^5.0.0",
+    "ed25519-hd-key": "^1.2.0",
+    "ethers": "^5.7.2",
+    "eventemitter3": "^5.0.1",
+    "tweetnacl": "^1.0.3"
   },
   "devDependencies": {
     "@types/eventemitter3": "^2.0.2",
